Added logging to file
This PR adds a new file appender to our log4j config dynamically when js2cpg is calles with `--log-file path/to/file`.
This requires to remove the `Runtime` flag at "org.apache.logging.log4j" % "log4j-slf4j-impl" in build.sbt as we need access to the actual logging implementation.

I do not really know if this breaks something @ release or in our CI/CD later on. Could we test this somehow?
Locally, everything is fine.