map JS comma operator to BLOCK nodes
TypeScript sometimes produces calls in the form
    (0, functionName)(params)

Translating this as previously to <operator>.commaright caused us to not
be able to resolve the callee to "functionName". This just works when we
use a BLOCK - which has the same semantics as the JS comma operator.

As always with AST changes there's some risk of breaking stuff.
I've tried to minimize this by using the same code field as previously
(even though it's invalid JavaScript in cases like this one).

Our test suites are happy with the change. There is one explicit
reference to the operator in CS that's easy enough to adapt.

We've also had a bunch of bugs over time where passes were doing string
matching on calls and ended up matching both the intended call and an
comma operator they were contained in - or were accidentally matching
unrelated arguments of the comma operator. Most of these should be fixed
by now, but I still feel this change fixes more such bugs than it
introduces in passes that actually try to deal with the comma operator.